requirements.txt
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
colorlog==6.6.0
google-auth==2.9.0
google-auth-oauthlib==0.5.2
gspread==5.4.0
gunicorn==20.1.0
idna==3.3
mutagen==1.45.1
numpy==1.23.0
oauthlib==3.2.0
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pytesseract==0.3.9
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
tqdm==4.64.0
urllib3==1.26.9
vidgear==0.2.5
websockets==10.3
yt-dlp==2022.6.29
